A target parser to be used with SFX OpenURL link resolver
There is a problem linking with SFX to the 500 restricted articles in
Harvard Business Review.  It seems that HBR is forcing EBSCO to block
SFX links to these articles because EBSCO is using persistent linking
behind the scenes to resolve the SFX OpenURLs. EBSCO has thus far
refused to lift
the restriction on "persistent links". The HBR parser is a workaround
that sends the article title and journal title to EBSCO as a search.